pyproject.toml
[tool.poetry]
name = "pythonproject"
version = "0.1.0"
description = ""
authors = ["jval7 <jhonvalderramaa7@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.13"
langchain = "^0.0.345"
openai = "^1.3.7"
boto3 = "^1.34.39"
pydantic = "^2.6.1"
black = "^24.1.1"
google-api-python-client = "^2.118.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.0"
requests = "^2.31.0"
google-auth = "^2.28.0"
pyjwt = "^2.8.0"
cryptography = "^42.0.3"
pytz = "^2024.1"
pytest = "^8.2.0"
coverage = "^7.5.1"
pytest-cov = "^5.0.0"
pymongo = "^4.7.2"
python-telegram-bot = "^21.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
